Add files via upload
Hey guys, i've been busy this week; however i am always into coding. Finally, i complete the DataBase with the items. In the Directory, the file ".img"  contains folder with the item's pics; they are about 90 items for this project . I also put some screenshot of this spint in the  "SCREENSHOT" folder. Enjoy.